+++ CHANGELOG.md
@@ -992,6 +992,8 @@
 - User deactivation via Devise
 - Two-factor authentication via SMS/OTP and a phased login including mobile number provision/confirmation
 - OTP resend for delivery failures
+- Relax user model validation to permit capitalisation changes and whitespace removal from email addresses
+- Use Devise case-insensitive method when retrieving users for authentication
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-99...HEAD
 [release-99]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-98...release-99